Split mazebuilder.builder in generator and type files
SRP at file level.
